sample/pyproject.toml
[project]
name = "sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.0",
    "cryptography>=41.0",
    "fastapi>=0.110.0",
    "gradio>=6.12.0",
    "jinja2>=3.1.0",
    "pandas>=2.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.27.0",
]
